requirement.txt
chainlit
python-dotenv
langchain
langchain-community
langchain-google-genai
chromadb
pypdf
PyMuPDF
Pillow